fix: add/subtract brightness using percent values
Since percentage is being calculated using human perception, the
relation between percentages and brightness is logarithmic and not
linear.
This means that operating on brightness values obtained from percent
values and converting them back to percents IS NOT the same as doing the
operations in percent values and then converting the result to
brightness values. The later is the preferred behaviour.